Replace placeholder slides with new images
These two slides should give me the introduction and the transition to
my actual talk. For the intro I'm focusing on the through-line of
TechPeople meetings -- connecting folks that are normally siloed. Then I
can transition in to my talk with a "And here's some tools you can use
to reduce these silos"