Draft 3
After finishing the skeleton I did a dry run and found some spots that
did not work. This led to a pass over nearly all of the slides.

A big part here is introducing a more standard structure to the 4 silo
sections.

- Introduce the silo
- Introduce problems due to that silo
- Introduce the tools we'll use to break that silo
- Then go tool/example, tool/example until done

More slides, but the structure is clearer and I think people will be
able to follow better

I dumped the wiki slide because I didn't find it useful. And good wikis
are super rare anyway.