Use big text, improve sub-step slides
I found that the "Use X to..." slides had a few problems

First, it could be hard to remember what pain that tool is meant to
solve. I've addressed this by adding a reminder of the problem.

Second, they were underwhelming visually. I'm now using Deckset's "big
text" settings to fix that.