Apply consistent styling across Silo sections
Going for a more uniform look across the 4 silo sections

A sore thumb was the Integrations slide. I've replaced the bulleted list
with a image, similar to other slides

Some images still had the footer, hiding it

Better use of big text styling in the What Next section, to make it more
consistent with the Silo sections.